# erasure-code/isa: Use isa/raid's xor_gen() instead of the region_xor(…