--- a/src/erasure-code/isa/ErasureCodeIsa.h
+++ b/src/erasure-code/isa/ErasureCodeIsa.h
@@ -30,6 +30,8 @@
 #include "ErasureCodeIsaTableCache.h"
 // -----------------------------------------------------------------------------
 
+#define EC_ISA_ADDRESS_ALIGNMENT 32u
+
 class ErasureCodeIsa : public ceph::ErasureCode {
 public: