… src/erasure-code/isa/xor_op.cc
… src/erasure-code/isa/xor_op.h