…) optimisation

When using the ISA plugin for erasure coded pools, we use an optimisation to simplify the xor encode and decode and improve performance. This optimisation applies when m=1 in the pool's erasure code profile (or if m in the profile is greater than 1 but the technique in use is reed_sol_van and the erasure is in a data chunk or the first parity chunk).

The current optimisation does improve decoding performance, but only takes advantage of SSE2 SIMD instructions on x86-64 CPUs. The ISA xor_gen() function in the ISA raid sub-directory can take advantage of newer SIMD instructions, and so using this results in a slight improvement to decoding performance when using ISA and the correct conditions to use the optimisation are met.

This pull request removes the region_xor code which is no longer needed, and instead uses the ISA raid xor_gen() function.

I have tested this change using the ISA unit tests

```
[root@533f6de09e3b build]# bin/unittest_erasure_code_isa
[==========] Running 10 tests from 1 test suite.
[----------] Global test environment set-up.
[----------] 10 tests from IsaErasureCodeTest
[ RUN      ] IsaErasureCodeTest.encode_decode
[       OK ] IsaErasureCodeTest.encode_decode (0 ms)
[ RUN      ] IsaErasureCodeTest.minimum_to_decode
[       OK ] IsaErasureCodeTest.minimum_to_decode (0 ms)
[ RUN      ] IsaErasureCodeTest.chunk_size
[       OK ] IsaErasureCodeTest.chunk_size (0 ms)
[ RUN      ] IsaErasureCodeTest.encode
[       OK ] IsaErasureCodeTest.encode (1 ms)
[ RUN      ] IsaErasureCodeTest.sanity_check_k
[       OK ] IsaErasureCodeTest.sanity_check_k (0 ms)
[ RUN      ] IsaErasureCodeTest.isa_vandermonde_exhaustive
[       OK ] IsaErasureCodeTest.isa_vandermonde_exhaustive (53 ms)
[ RUN      ] IsaErasureCodeTest.isa_cauchy_exhaustive
[       OK ] IsaErasureCodeTest.isa_cauchy_exhaustive (53 ms)
[ RUN      ] IsaErasureCodeTest.isa_cauchy_cache_trash
[       OK ] IsaErasureCodeTest.isa_cauchy_cache_trash (217 ms)
[ RUN      ] IsaErasureCodeTest.isa_xor_codec
[       OK ] IsaErasureCodeTest.isa_xor_codec (1 ms)
[ RUN      ] IsaErasureCodeTest.create_rule
[       OK ] IsaErasureCodeTest.create_rule (0 ms)
[----------] 10 tests from IsaErasureCodeTest (326 ms total)

[----------] Global test environment tear-down
[==========] 10 tests from 1 test suite ran. (326 ms total)
[  PASSED  ] 10 tests.
```

And used the ceph erasure coding benchmark on an Intel Xeon Gold 6336Y CPU (TOTAL_SIZE=$((4 * 1024 * 1024 * 1024)) erasure-code/bench.sh fplot | tee erasure-code/bench.js).
Results show very similar performance encoding, but a minor improvement when decoding and M = 1, or number of erasures = 1

```
baseline, without this PR
Encode reed_sol_van
Plugin	Technique	Time	Total Size	k	m	Iteration	Packet Size
isa	reed_sol_van	0.810144	4194304	2	1	1048576	256
isa	reed_sol_van	1.239091	4194304	3	2	1048576	160
isa	reed_sol_van	1.070481	4194304	4	2	1048576	128
isa	reed_sol_van	1.491363	4194304	4	3	1048576	128
isa	reed_sol_van	1.560132	4194304	6	2	1048576	80
isa	reed_sol_van	1.819574	4194304	6	3	1048576	80
isa	reed_sol_van	2.21709	4194304	6	4	1048576	80
isa	reed_sol_van	2.356909	4194304	10	3	1048576	48
isa	reed_sol_van	2.679524	4194304	10	4	1048576	48

with this PR
Encode reed_sol_van
Plugin	Technique	Time	Total Size	k	m	Iteration	Packet Size
isa	reed_sol_van	0.819062	4194304	2	1	1048576	256
isa	reed_sol_van	1.239847	4194304	3	2	1048576	160
isa	reed_sol_van	1.05034	4194304	4	2	1048576	128
isa	reed_sol_van	1.492579	4194304	4	3	1048576	128
isa	reed_sol_van	1.546993	4194304	6	2	1048576	80
isa	reed_sol_van	1.795169	4194304	6	3	1048576	80
isa	reed_sol_van	2.198842	4194304	6	4	1048576	80
isa	reed_sol_van	2.340041	4194304	10	3	1048576	48
isa	reed_sol_van	2.718091	4194304	10	4	1048576	48

baseline, without this PR
Encode cauchy
Plugin	Technique	Time	Total Size	k	m	Iteration	Packet Size
isa	cauchy	0.823824	4194304	2	1	1048576	256
isa	cauchy	1.263262	4194304	3	2	1048576	160
isa	cauchy	1.062328	4194304	4	2	1048576	128
isa	cauchy	1.517931	4194304	4	3	1048576	128
isa	cauchy	1.5555	4194304	6	2	1048576	80
isa	cauchy	1.804116	4194304	6	3	1048576	80
isa	cauchy	2.199236	4194304	6	4	1048576	80
isa	cauchy	2.338742	4194304	10	3	1048576	48
isa	cauchy	2.780651	4194304	10	4	1048576	48

with this PR
Encode cauchy
Plugin	Technique	Time	Total Size	k	m	Iteration	Packet Size
isa	cauchy	0.824136	4194304	2	1	1048576	256
isa	cauchy	1.211994	4194304	3	2	1048576	160
isa	cauchy	1.045071	4194304	4	2	1048576	128
isa	cauchy	1.494725	4194304	4	3	1048576	128
isa	cauchy	1.53728	4194304	6	2	1048576	80
isa	cauchy	1.784101	4194304	6	3	1048576	80
isa	cauchy	2.221333	4194304	6	4	1048576	80
isa	cauchy	2.33441	4194304	10	3	1048576	48
isa	cauchy	2.652096	4194304	10	4	1048576	48

baseline, without this PR
Decode reed_sol_van
Plugin	Technique	Time	Total Size	k	m	Iteration	Packet Size	Erasures
isa	reed_sol_van	0.899744	4194304	2	1	1048576	256	1
isa	reed_sol_van	1.724666	4194304	3	2	1048576	160	1
isa	reed_sol_van	2.314405	4194304	3	2	1048576	160	2
isa	reed_sol_van	1.920687	4194304	4	2	1048576	128	1
isa	reed_sol_van	2.726822	4194304	4	2	1048576	128	2
isa	reed_sol_van	2.110076	4194304	4	3	1048576	128	1
isa	reed_sol_van	2.825197	4194304	4	3	1048576	128	2
isa	reed_sol_van	3.418693	4194304	4	3	1048576	128	3
isa	reed_sol_van	2.331806	4194304	6	2	1048576	80	1
isa	reed_sol_van	3.413767	4194304	6	2	1048576	80	2
isa	reed_sol_van	2.61269	4194304	6	3	1048576	80	1
isa	reed_sol_van	3.637135	4194304	6	3	1048576	80	2
isa	reed_sol_van	4.215466	4194304	6	3	1048576	80	3
isa	reed_sol_van	2.876675	4194304	6	4	1048576	80	1
isa	reed_sol_van	3.881037	4194304	6	4	1048576	80	2
isa	reed_sol_van	4.636734	4194304	6	4	1048576	80	3
isa	reed_sol_van	5.281095	4194304	6	4	1048576	80	4
isa	reed_sol_van	3.657244	4194304	10	3	1048576	48	1
isa	reed_sol_van	4.894165	4194304	10	3	1048576	48	2
isa	reed_sol_van	5.783632	4194304	10	3	1048576	48	3
isa	reed_sol_van	4.099728	4194304	10	4	1048576	48	1
isa	reed_sol_van	5.413802	4194304	10	4	1048576	48	2
isa	reed_sol_van	6.250787	4194304	10	4	1048576	48	3
isa	reed_sol_van	7.005922	4194304	10	4	1048576	48	4

with this PR
Decode reed_sol_van
Plugin	Technique	Time	Total Size	k	m	Iteration	Packet Size	Erasures
isa	reed_sol_van	0.885898	4194304	2	1	1048576	256	1
isa	reed_sol_van	1.638085	4194304	3	2	1048576	160	1
isa	reed_sol_van	2.302119	4194304	3	2	1048576	160	2
isa	reed_sol_van	1.871081	4194304	4	2	1048576	128	1
isa	reed_sol_van	2.72915	4194304	4	2	1048576	128	2
isa	reed_sol_van	2.052196	4194304	4	3	1048576	128	1
isa	reed_sol_van	2.836173	4194304	4	3	1048576	128	2
isa	reed_sol_van	3.340072	4194304	4	3	1048576	128	3
isa	reed_sol_van	2.301173	4194304	6	2	1048576	80	1
isa	reed_sol_van	3.440601	4194304	6	2	1048576	80	2
isa	reed_sol_van	2.573077	4194304	6	3	1048576	80	1
isa	reed_sol_van	3.647883	4194304	6	3	1048576	80	2
isa	reed_sol_van	4.226653	4194304	6	3	1048576	80	3
isa	reed_sol_van	2.841784	4194304	6	4	1048576	80	1
isa	reed_sol_van	3.871163	4194304	6	4	1048576	80	2
isa	reed_sol_van	4.646203	4194304	6	4	1048576	80	3
isa	reed_sol_van	5.235489	4194304	6	4	1048576	80	4
isa	reed_sol_van	3.407821	4194304	10	3	1048576	48	1
isa	reed_sol_van	4.936023	4194304	10	3	1048576	48	2
isa	reed_sol_van	5.77874	4194304	10	3	1048576	48	3
isa	reed_sol_van	3.81006	4194304	10	4	1048576	48	1
isa	reed_sol_van	5.40965	4194304	10	4	1048576	48	2
isa	reed_sol_van	6.272648	4194304	10	4	1048576	48	3
isa	reed_sol_van	7.001062	4194304	10	4	1048576	48	4

baseline, without this PR
Decode cauchy
Plugin	Technique	Time	Total Size	k	m	Iteration	Packet Size	Erasures
isa	cauchy	0.905557	4194304	2	1	1048576	256	1
isa	cauchy	1.889728	4194304	3	2	1048576	160	1
isa	cauchy	2.308956	4194304	3	2	1048576	160	2
isa	cauchy	2.352925	4194304	4	2	1048576	128	1
isa	cauchy	2.733845	4194304	4	2	1048576	128	2
isa	cauchy	2.436355	4194304	4	3	1048576	128	1
isa	cauchy	2.825513	4194304	4	3	1048576	128	2
isa	cauchy	3.347051	4194304	4	3	1048576	128	3
isa	cauchy	2.87374	4194304	6	2	1048576	80	1
isa	cauchy	3.428421	4194304	6	2	1048576	80	2
isa	cauchy	3.141515	4194304	6	3	1048576	80	1
isa	cauchy	3.65206	4194304	6	3	1048576	80	2
isa	cauchy	4.21947	4194304	6	3	1048576	80	3
isa	cauchy	3.341173	4194304	6	4	1048576	80	1
isa	cauchy	3.89945	4194304	6	4	1048576	80	2
isa	cauchy	4.627455	4194304	6	4	1048576	80	3
isa	cauchy	5.234281	4194304	6	4	1048576	80	4
isa	cauchy	4.465539	4194304	10	3	1048576	48	1
isa	cauchy	4.911119	4194304	10	3	1048576	48	2
isa	cauchy	5.784103	4194304	10	3	1048576	48	3
isa	cauchy	4.642966	4194304	10	4	1048576	48	1
isa	cauchy	5.408909	4194304	10	4	1048576	48	2
isa	cauchy	6.407602	4194304	10	4	1048576	48	3
isa	cauchy	7.033372	4194304	10	4	1048576	48	4

with this PR
Decode cauchy
Plugin	Technique	Time	Total Size	k	m	Iteration	Packet Size	Erasures
isa	cauchy	0.883152	4194304	2	1	1048576	256	1
isa	cauchy	1.90028	4194304	3	2	1048576	160	1
isa	cauchy	2.316268	4194304	3	2	1048576	160	2
isa	cauchy	2.235922	4194304	4	2	1048576	128	1
isa	cauchy	2.817762	4194304	4	2	1048576	128	2
isa	cauchy	2.445407	4194304	4	3	1048576	128	1
isa	cauchy	2.839687	4194304	4	3	1048576	128	2
isa	cauchy	3.335637	4194304	4	3	1048576	128	3
isa	cauchy	2.866535	4194304	6	2	1048576	80	1
isa	cauchy	3.417326	4194304	6	2	1048576	80	2
isa	cauchy	3.128992	4194304	6	3	1048576	80	1
isa	cauchy	3.672589	4194304	6	3	1048576	80	2
isa	cauchy	4.229856	4194304	6	3	1048576	80	3
isa	cauchy	3.315932	4194304	6	4	1048576	80	1
isa	cauchy	3.881264	4194304	6	4	1048576	80	2
isa	cauchy	4.661972	4194304	6	4	1048576	80	3
isa	cauchy	5.298398	4194304	6	4	1048576	80	4
isa	cauchy	4.425217	4194304	10	3	1048576	48	1
isa	cauchy	4.934854	4194304	10	3	1048576	48	2
isa	cauchy	5.881886	4194304	10	3	1048576	48	3
isa	cauchy	4.67152	4194304	10	4	1048576	48	1
isa	cauchy	5.547466	4194304	10	4	1048576	48	2
isa	cauchy	6.277136	4194304	10	4	1048576	48	3
isa	cauchy	7.039528	4194304	10	4	1048576	48	4
```


<!--
  - Please give your pull request a title like

      [component]: [short description]

  - Please use this format for each git commit message:

      [component]: [short description]

      [A longer multiline description]

      Fixes: [ticket URL on tracker.ceph.com, create one if necessary]
      Signed-off-by: [Your Name] <[your email]>

    For examples, use "git log".
-->

## Contribution Guidelines
- To sign and title your commits, please refer to [Submitting Patches to Ceph](https://github.com/ceph/ceph/blob/main/SubmittingPatches.rst).

- If you are submitting a fix for a stable branch (e.g. "quincy"), please refer to [Submitting Patches to Ceph - Backports](https://github.com/ceph/ceph/blob/master/SubmittingPatches-backports.rst) for the proper workflow.

- When filling out the below checklist, you may click boxes directly in the GitHub web UI.  When entering or editing the entire PR message in the GitHub web UI editor, you may also select a checklist item by adding an `x` between the brackets: `[x]`.  Spaces and capitalization matter when checking off items this way.

## Checklist
- Tracker (select at least one)
  - [ ] References tracker ticket
  - [ ] Very recent bug; references commit where it was introduced
  - [x] New feature (ticket optional)
  - [ ] Doc update (no ticket needed)
  - [ ] Code cleanup (no ticket needed)
- Component impact
  - [ ] Affects [Dashboard](https://tracker.ceph.com/projects/dashboard/issues/new), opened tracker ticket
  - [ ] Affects [Orchestrator](https://tracker.ceph.com/projects/orchestrator/issues/new), opened tracker ticket
  - [x] No impact that needs to be tracked
- Documentation (select at least one)
  - [ ] Updates relevant documentation
  - [x] No doc update is appropriate
- Tests (select at least one)
  - [ ] Includes [unit test(s)](https://docs.ceph.com/en/latest/dev/developer_guide/tests-unit-tests/)
  - [ ] Includes [integration test(s)](https://docs.ceph.com/en/latest/dev/developer_guide/testing_integration_tests/)
  - [ ] Includes bug reproducer
  - [x] No tests

<details>
<summary>Show available Jenkins commands</summary>

- `jenkins retest this please`
- `jenkins test classic perf`
- `jenkins test crimson perf`
- `jenkins test signed`
- `jenkins test make check`
- `jenkins test make check arm64`
- `jenkins test submodules`
- `jenkins test dashboard`
- `jenkins test dashboard cephadm`
- `jenkins test api`
- `jenkins test docs`
- `jenkins render docs`
- `jenkins test ceph-volume all`
- `jenkins test ceph-volume tox`
- `jenkins test windows`
- `jenkins test rook e2e`
</details>