@jamiepryde 2024-07-24
Some additional data from the benchmark running in a CentOS 9 container on a macbook pro with an M1 pro CPU (ARM)

```
Encode
Plugin	Technique	Baseline Time	New Time	Total Size	k	m	Iteration	Packet Size
isa	reed_sol_van	0.529021	0.484336	4194304		2	1	1048576		256
isa	reed_sol_van	1.892235	1.914296	4194304		3	2	1048576		160
isa	reed_sol_van	1.818443	1.903929	4194304		4	2	1048576		128
isa	reed_sol_van	2.302543	2.015336	4194304		4	3	1048576		128
isa	reed_sol_van	2.487124	2.274934	4194304		6	2	1048576		80
isa	reed_sol_van	2.432974	2.424385	4194304		6	3	1048576		80
isa	reed_sol_van	3.048638	2.830223	4194304		6	4	1048576		80
isa	reed_sol_van	3.1322100	3.167969	4194304		10	3	1048576		48
isa	reed_sol_van	3.606801	3.634400	4194304		10	4	1048576		48

isa	cauchy		0.535701	0.482847	4194304		2	1	1048576		256
isa	cauchy		1.984469	1.918598	4194304		3	2	1048576		160
isa	cauchy		1.736924	1.775383	4194304		4	2	1048576		128
isa	cauchy		2.063875	2.000363	4194304		4	3	1048576		128
isa	cauchy		2.301490	2.285416	4194304		6	2	1048576		80
isa	cauchy		2.407579	2.373702	4194304		6	3	1048576		80
isa	cauchy		2.791605	2.841477	4194304		6	4	1048576		80
isa	cauchy		3.0511270	2.993474	4194304		10	3	1048576		48
isa	cauchy		3.562984	3.597210	4194304		10	4	1048576		48

Decode

Plugin	Technique	Baseline Time	New Time	Total Size	k	m	Iteration	Packet Size	Erasures
isa	reed_sol_van	0.824144	0.700814	4194304		2	1	1048576		256		1
isa	reed_sol_van	1.459985	1.389423	4194304		3	2	1048576		160		1
isa	reed_sol_van	2.7416510	2.700939	4194304		3	2	1048576		160		2
isa	reed_sol_van	1.6147340	1.539243	4194304		4	2	1048576		128		1
isa	reed_sol_van	2.986448	2.857361	4194304		4	2	1048576		128		2
isa	reed_sol_van	1.876038	1.882001	4194304		4	3	1048576		128		1
isa	reed_sol_van	3.062447	3.084523	4194304		4	3	1048576		128		2
isa	reed_sol_van	3.434438	3.301625	4194304		4	3	1048576		128		3
isa	reed_sol_van	2.022273	1.910465	4194304		6	2	1048576		80		1
isa	reed_sol_van	3.789703	3.606070	4194304		6	2	1048576		80		2
isa	reed_sol_van	2.421964	2.269290	4194304		6	3	1048576		80		1
isa	reed_sol_van	3.877603	3.774262	4194304		6	3	1048576		80		2
isa	reed_sol_van	4.062006	3.983650	4194304		6	3	1048576		80		3
isa	reed_sol_van	2.588132	2.6101130	4194304		6	4	1048576		80		1
isa	reed_sol_van	4.044973	4.044441	4194304		6	4	1048576		80		2
isa	reed_sol_van	4.281397	4.374214	4194304		6	4	1048576		80		3
isa	reed_sol_van	4.944261	4.867536	4194304		6	4	1048576		80		4
isa	reed_sol_van	3.047575	3.052979	4194304		10	3	1048576		48		1
isa	reed_sol_van	5.031783	4.9418110	4194304		10	3	1048576		48		2
isa	reed_sol_van	5.643340	5.636444	4194304		10	3	1048576		48		3
isa	reed_sol_van	3.446319	3.344000	4194304		10	4	1048576		48		1
isa	reed_sol_van	5.289960	5.313369	4194304		10	4	1048576		48		2
isa	reed_sol_van	5.989640	5.966225	4194304		10	4	1048576		48		3
isa	reed_sol_van	6.874030	6.725350	4194304		10	4	1048576		48		4

isa	cauchy		0.774834	0.696212	4194304		2	1	1048576		256		1
isa	cauchy		2.257658	2.260327	4194304		3	2	1048576		160		1
isa	cauchy		2.7141160	2.7527410	4194304		3	2	1048576		160		2
isa	cauchy		2.473062	2.503689	4194304		4	2	1048576		128		1
isa	cauchy		2.912288	2.976777	4194304		4	2	1048576		128		2
isa	cauchy		2.712970	2.660301	4194304		4	3	1048576		128		1
isa	cauchy		3.027421	3.093792	4194304		4	3	1048576		128		2
isa	cauchy		3.365696	3.416050	4194304		4	3	1048576		128		3
isa	cauchy		3.140242	3.075421	4194304		6	2	1048576		80		1
isa	cauchy		3.679346	3.624555	4194304		6	2	1048576		80		2
isa	cauchy		3.516186	3.265223	4194304		6	3	1048576		80		1
isa	cauchy		4.068511	3.796682	4194304		6	3	1048576		80		2
isa	cauchy		4.453731	3.980017	4194304		6	3	1048576		80		3
isa	cauchy		3.543436	3.526323	4194304		6	4	1048576		80		1
isa	cauchy		4.156210	4.235436	4194304		6	4	1048576		80		2
isa	cauchy		4.514992	4.378862	4194304		6	4	1048576		80		3
isa	cauchy		5.129092	5.1993710	4194304		6	4	1048576		80		4
isa	cauchy		4.392270	4.401733	4194304		10	3	1048576		48		1
isa	cauchy		5.080615	4.974371	4194304		10	3	1048576		48		2
isa	cauchy		5.750088	5.667986	4194304		10	3	1048576		48		3
isa	cauchy		4.797167	4.646359	4194304		10	4	1048576		48		1
isa	cauchy		5.424269	5.289619	4194304		10	4	1048576		48		2
isa	cauchy		6.086570	5.776426	4194304		10	4	1048576		48		3
isa	cauchy		6.831321	6.945226	4194304		10	4	1048576		48		4

```